requirements.txt
openai>=1.0
opencv-python>=4.8
pillow
watchdog
pdfkit
jinja2
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
